# openapiDoc.JSON
{
  "openapi": "3.0.3",
  "info": {
    "title": "Groupomania social network",
    "description": "This is the documentation of Groupomania social network API.",
    "termsOfService": "http://groupomania.com/terms/",
    "contact": {
      "name": "API support",
      "url": "http://groupomania.com/contact/",
      "email": "contactAPI@groupomania.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [{
    "url": "http://localhost:{port}/{basePath}",
    "variables": {
      "port": {
        "default": "3000"
      },
      "basePath": {
        "default": "api"
      }
    }
  }],
...
  "paths": {
...
    "/posts": {
      "post": {
        "tags": ["posts"],
        "summary": "Create a post.",
        "description": "User creates a post. Form-data is required if an image file is added. File type must be 'image/jpeg', 'image/jpg' or 'image/png'.",
        "operationId": "createOnePost",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "Post properties. For more details, go to 'PostJson' or 'PostFormdata' schemas.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/postJson"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/postFormdata"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Post created."
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Unauthorized request."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "get": {
        "tags": ["posts"],
        "summary": "Get all posts.",
        "description": "User gets all posts.",
        "operationId": "getAllPosts",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          },
          {
            "name": "minDate",
            "in": "query",
            "description": "Get all posts who have been created after this date. Only the ISO 8601 format (YYYY-MM-DDTHH:mm:ss.sssZ) is supported.",
            "schema": {
              "type": "string",
              "example": "2022-13-01T14:48:02.809Z"
            }
          },
          {
            "name": "maxDate",
            "in": "query",
            "description": "Get all posts who have been created before this date. Only the ISO 8601 format (YYYY-MM-DDTHH:mm:ss.sssZ) is supported.",
            "schema": {
              "type": "string",
              "example": "2022-13-01T14:48:02.809Z"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Get a limited amount of posts.",
            "schema": {
              "type": "number",
              "example": 3
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Returns elements from oldest to newest ('asc') or from newest to oldest ('desc').",
            "schema": {
              "type": "string",
              "example": "asc"
            }
          },
          {
            "name": "fromUserId",
            "in": "query",
            "description": "Get all posts from a specific user.",
            "schema": {
              "type": "string",
              "example": "632dc722c75414645a459eb2"
            }
          },
          {
            "name": "reactions",
            "in": "query",
            "description": "Get reactions of the posts.",
            "schema": {
              "type": "boolean",
              "example": true
            }
          },
          {
            "name": "comments",
            "in": "query",
            "description": "Get comments of the posts.",
            "schema": {
              "type": "boolean",
              "example": true
            }
          },
          {
            "name": "commentsReactions",
            "in": "query",
            "description": "Get reactions of the posts comments. Only available if comments query is true.",
            "schema": {
              "type": "boolean",
              "example": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/post"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized request."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "delete": {
        "tags": ["posts"],
        "summary": "Delete all posts.",
        "description": "Admin deletes all posts.",
        "operationId": "deleteAllPosts",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token. Must be an admin one.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Ok."
          },
          "403": {
            "description": "Forbidden request."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/posts/{postId}": {
      "get": {
        "tags": ["posts"],
        "summary": "Get one post.",
        "description": "User gets properties of a post.",
        "operationId": "getOnePost",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          },
          {
            "name": "postId",
            "in": "path",
            "description": "Id of post to return.",
            "schema": {
              "type": "string",
              "example": "632efd854148fdc69cd7e652"
            },
            "required": true
          },
          {
            "name": "reactions",
            "in": "query",
            "description": "Get reactions of the post.",
            "schema": {
              "type": "boolean",
              "example": true
            }
          },
          {
            "name": "comments",
            "in": "query",
            "description": "Get comments of the post.",
            "schema": {
              "type": "boolean",
              "example": true
            }
          },
          {
            "name": "commentsReactions",
            "in": "query",
            "description": "Get reactions of the post comments.",
            "schema": {
              "type": "boolean",
              "example": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/post"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized request."
          },
          "404": {
            "description": "Invalid Id."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "put": {
        "tags": ["posts"],
        "summary": "Modify one post.",
        "description": "User modifies a post of his own, or admin modifies a post. Form-data is required if an image file is added. File type must be 'image/jpeg', 'image/jpg' or 'image/png'.",
        "operationId": "modifyOnePost",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          },
          {
            "name": "postId",
            "in": "path",
            "description": "Id of post to modify.",
            "schema": {
              "type": "string",
              "example": "632efd854148fdc69cd7e652"
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "Post properties modifications. For more details, go to 'PostModifyJson' or 'PostModifyFormData' schemas. Post must always have at least a content or an image.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/postModifyJson"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/postModifyFormData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Ok."
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Unauthorized request."
          },
          "404": {
            "description": "Invalid Id."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "delete": {
        "tags": ["posts"],
        "summary": "Delete one post.",
        "description": "User deletes a post of his own, or admin deletes a post.",
        "operationId": "deleteOnePost",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          },
          {
            "name": "postId",
            "in": "path",
            "description": "Id of post to delete.",
            "schema": {
              "type": "string",
              "example": "632efd854148fdc69cd7e652"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Ok."
          },
          "401": {
            "description": "Unauthorized request."
          },
          "404": {
            "description": "Invalid Id."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "post": {
        "title": "Post",
        "description": "Post schema. At least one of 'content' or 'imageUrl' is required.",
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "632efd854148fdc69cd7e652"
          },
          "content": {
            "type": "string",
            "example": "Voici une image."
          },
          "imageUrl": {
            "type": "string",
            "example": "images/monImage.jpg"
          },
          "userId": {
            "type": "string",
            "example": "632dc722c75414645a459eb2"
          },
          "creationDate": {
            "type": "string",
            "example": "2022-13-01T14:48:02.809Z"
          },
          "reactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/reaction"
            }
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/comment"
            }
          },
          "__v": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "_id",
          "userId",
          "creationDate",
          "__v"
        ]
      },
...
      "reaction": {
        "title": "Reaction",
        "description": "Reaction schema. Exactly one of 'postId' or 'commentId' must be completed.",
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "632efdbb4148fdc69cd7e784"
          },
          "type": {
            "type": "string",
            "example": "Like"
          },
          "postId": {
            "type": "string",
            "example": ""
          },
          "commentId": {
            "type": "string",
            "example": "632efdbb4148fdc69cd7e654"
          },
          "userId": {
            "type": "string",
            "example": "632dc722c75414645a459eb2"
          },
          "creationDate": {
            "type": "string",
            "example": "2022-13-01T14:48:02.809Z"
          },
          "__v": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "_id",
          "type",
          "userId",
          "creationDate",
          "__v"
        ]
      },
...
      "bearerToken": {
        "description": "Bearer token schema.",
        "type": "string",
        "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MzMxZDk1MWU3N2YxNDAxMmQxMjExNmEiLCJpc0FkbWluIjpmYWxzZSwiaWF0IjoxNjY0MjExMzEyLCJleHAiOjE2NjQyOTc3MTJ9.03h-xFcgHmwpYCHMZq36hnSAsHNS7TItboYmAliKSIk"
      },
...
      "postJson": {
        "title": "PostJson",
        "description": "Post schema used when user publishes or modifies his post without any image.",
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "Voici une image."
          }
        },
        "required": [
        "content"
        ]
      },
      "postModifyJson": {
        "title": "PostModifyJson",
        "description": "Post schema used when user modifies his post without any image. He can delete post image with deletePostImage property and post content with deletepostContent property. Post must always have at least a content or an image.",
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "Voici une image."
          },
          "deletePostImage": {
            "type": "boolean",
            "example": true
          },
          "deletePostContent": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "postFormdata": {
        "title": "PostFormdata",
        "description": "Post schema used when user publishes/modifies his post or admin modifies a post with an image file. File type must be 'image/jpeg', 'image/jpg' or 'image/png'.",
        "type": "object",
        "properties": {
          "post": {
            "type": "object",
            "properties": {
              "content": {
                "type": "string",
                "example": "Voici une image."
               }
            }
          },
          "image": {
            "type": "string",
            "example": "MonImage.jpg"
          }
        },
        "required": [
          "image"
        ]
      },
      "postModifyFormData": {
        "title": "PostModifyFormData",
        "description": "Post schema used when user modifies his post with an image. He can delete post content with deletepostContent property. Post must always have at least a content or an image.",
         "type": "object",
        "properties": {
          "post": {
            "type": "object",
            "properties": {
              "content": {
                "type": "string",
                "example": "Voici une image."
              },
              "deletePostContent": {
                "type": "boolean",
                "example": false
              }
            }
          },
          "image": {
            "type": "string",
            "example": "MonImage.jpg"
          }
        },
        "required": [
          "image"
        ]
      },
...
    }
  }
}